pyproject.toml
[project]
name = "olashlibraryuv"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "mysqlclient>=2.2.7",
    "python-dotenv>=1.1.0",
]
